+++ server/src/main/java/com/example/demo/repository/CapacitorUnitRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.CapacitorUnit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CapacitorUnitRepository extends JpaRepository<CapacitorUnit, Integer> {
+public interface CapacitorUnitRepository extends JpaRepository<CapacitorUnit, Long> {
 }
